Transaction still has to be linked to the database for GUI to update. Import transaction button prompts for file path, need to add the import function to action listener.

+++ tst/ImportTransactionTest.java
@@ -26,7 +26,7 @@ public class ImportTransactionTest {
 		Integer accountID = 1;
 		String type = "deposit";
 		String date = "09-09-1999";
-		Float amount = 100.0f;
+		Integer amount = 100;
 		expected = new TransactionModel();
 		expected.setAccountId(accountID);
 		expected.setType(type);
