+++ tst/TransactionRepositoryTest.java
@@ -23,7 +23,7 @@ public class TransactionRepositoryTest {
 		Integer accountID = 1;
 		String type = "deposit";
 		String date = "09-09-1999";
-		Float amount = 100.0f;
+		Integer amount = 100;
 		TransactionModel expected = new TransactionModel();
 		expected.setAccountId(accountID);
 		expected.setType(type);
